Stack nav completion on mount
StackView is responsible for calling the navigation completion action when `state.isTransitioning` is set. This fix handles that case when the stack is first mounting.

This had caused the following issue when observing the events.. because isTransitioning was never switched to false, didFocus never gets called: https://github.com/react-navigation/react-navigation/issues/4867