fix: attempt to fix screens integration
This seems to work on both Android and iOS. However there's still one issue on iOS.
When we call `setNativeProps`, it seems to interrupt the current gesture.
Since we call it to make the previous screen active on gesture start, it immediately ends the gesture.

Another issue is that the modal animation doesn't work great when screens is enabled.
since the previous screen needs to stay visible for the iOS modal presentation animation.