feat: change header height
Often times we want to change the header height on a per screen basis.
However, a limitation exists with the current method of changing the
height via `headerStyle`. Using `headerStyle.height` results in a header
that centers everything vertically. A common pattern in iOS is to have
the back button with what essentially amounts to a `top: 0` value.

This PR adds the `headerHeight` option to circumvent this.